+++ hermes2d/include/quadrature/limit_order.h
@@ -22,11 +22,11 @@ namespace Hermes
   namespace Hermes2D
   {
     /// can be called to set a custom order limiting table
-    extern HERMES_API void set_order_limit_table(int* tri_table, int* quad_table, int n);
+    extern HERMES_API void set_order_limit_table(int* tri_table, unsigned short* quad_table, unsigned short n);
 
     /// limit_order is used in integrals
-    extern HERMES_API int  g_safe_max_order;
-    extern HERMES_API int  g_max_order;
+    extern HERMES_API unsigned short  g_safe_max_order;
+    extern HERMES_API unsigned short  g_max_order;
 
     extern HERMES_API void update_limit_table(ElementMode2D mode);
     extern HERMES_API void limit_order(int& o, ElementMode2D mode);
