+++ hermes2d/include/shapeset/shapeset_h1_all.h
@@ -31,7 +31,7 @@ namespace Hermes
       H1ShapesetJacobi();
       virtual Shapeset* clone() { return new H1ShapesetJacobi(*this); };
       virtual SpaceType get_space_type() const { return HERMES_H1_SPACE; }
-      virtual int get_max_index(ElementMode2D mode) const;
+      virtual unsigned short get_max_index(ElementMode2D mode) const;
       
       virtual unsigned char get_id() const { return 1; }
       
@@ -46,14 +46,14 @@ namespace Hermes
     public:
       H1ShapesetEigen();
       virtual Shapeset* clone() { return new H1ShapesetEigen(*this); };
-      virtual int get_max_index(ElementMode2D mode) {};
+      virtual unsigned short get_max_index(ElementMode2D mode) {};
     private:
       virtual unsigned char get_id() const { return 2; }
       template<typename Scalar> friend class VectorForm;
       template<typename Scalar> friend class MatrixForm;
       virtual SpaceType get_space_type() const { return HERMES_H1_SPACE; }
       template<typename Scalar> friend class DiscreteProblem; template<typename Scalar> friend class Solution; friend class CurvMap; friend class RefMap; template<typename Scalar> friend class RefinementSelectors::H1ProjBasedSelector; template<typename Scalar> friend class RefinementSelectors::L2ProjBasedSelector; template<typename Scalar> friend class RefinementSelectors::HcurlProjBasedSelector; template<typename Scalar> friend class RefinementSelectors::OptimumSelector; friend class PrecalcShapeset;
-      static const int max_index[H2D_NUM_MODES];
+      static const unsigned short max_index[H2D_NUM_MODES];
     };
     */
