+++ hermes2d/include/shapeset/shapeset_hc_all.h
@@ -29,11 +29,11 @@ namespace Hermes
     public:
       HcurlShapesetGradLeg();
       virtual Shapeset* clone() { return new HcurlShapesetGradLeg(*this); };
-      virtual int get_max_index(ElementMode2D mode) const;
+      virtual unsigned short get_max_index(ElementMode2D mode) const;
       virtual unsigned char get_id() const { return 13; }
       virtual SpaceType get_space_type() const { return HERMES_HCURL_SPACE; }
       
-      static const int max_index[H2D_NUM_MODES];
+      static const unsigned short max_index[H2D_NUM_MODES];
     };
 
     /// This is the default Hcurl shapeset typedef.
