+++ hermes2d/include/shapeset/shapeset_hd_all.h
@@ -30,10 +30,10 @@ namespace Hermes
       HdivShapesetLegendre();
       virtual Shapeset* clone() { return new HdivShapesetLegendre(*this); };
       virtual SpaceType get_space_type() const { return HERMES_HDIV_SPACE; }
-      virtual int get_max_index(ElementMode2D mode) const;
+      virtual unsigned short get_max_index(ElementMode2D mode) const;
       virtual unsigned char get_id() const { return 20; }
       
-      static const int max_index[H2D_NUM_MODES];
+      static const unsigned short max_index[H2D_NUM_MODES];
     };
 
     /// This is the default Hdiv shapeset typedef.
