+++ hermes2d/include/views/thread_linearizer.h
@@ -90,7 +90,7 @@ namespace Hermes
         /// quad:: split == 1 -> horizontal split.
         /// quad:: split == 2 -> vertical split.
         /// quad:: split == 3 -> split to four quads.
-        void split_decision(int& split, int iv0, int iv1, int iv2, int iv3, ElementMode2D mode, const double** val, double* phx, double* phy, int* indices) const;
+        void split_decision(int& split, int iv0, int iv1, int iv2, int iv3, ElementMode2D mode, const double** val, double* phx, double* phy, unsigned short* indices) const;
 
         /// Utility - check of the orientation of tthe two triangles outputted for a quad.
         bool quad_flip(int iv0, int iv1, int iv2, int iv3) const;
